remove dtend from th xCal export
xCal is a direct port of iCal to XML.
iCal only allows one of dtend and duration to be specified.
The non-normative relax-ng schema in the relevant rfc reflects this.

I removed the duration and not the dtend, as parsers have to support the
datetime format anyways and there are parsers without support for the
duration format.

xCal [rfc 6321](https://tools.ietf.org/html/rfc6321#page-32):

    ```
    (property-dtend | property-duration)? &
    ```

iCal [rfc 5545](https://tools.ietf.org/html/rfc5545#page52):

    ```
    ; Either 'dtend' or 'duration' MAY appear in
    ; a 'eventprop', but 'dtend' and 'duration'
    ; MUST NOT occur in the same 'eventprop'.
    ;
    dtend / duration /
    ```